#361 - Move __debugInfo() to  ComPagesObjectDebuggable
Improve the __debugInfo() implementation making it re-usable
and and transform the properties recursively
