+++ code/site/components/com_pages/model/entity/item.php
@@ -9,6 +9,8 @@
 
 class ComPagesModelEntityItem extends KModelEntityAbstract implements ComPagesModelEntityInterface
 {
+    use ComPagesObjectDebuggable;
+
     private $__internal_properties;
 
     public function __construct(KObjectConfig $config)
@@ -70,56 +72,20 @@ class ComPagesModelEntityItem extends KModelEntityAbstract implements ComPagesMo
             $data[$property] = $this->{$property};
         }
 
-        foreach($data as $key => $value)
-        {
-            //Unpack config objects
-            if($value instanceof KObjectConfigInterface) {
-                $data[$key] = KObjectConfig::unbox($value);
-            }
-
-            //Remove NULL values
-            if(is_null($value)) {
-                unset($data[$key]);
-            }
+        //Add none-internal computed properties
+        foreach(array_diff($computed, $internal) as $property) {
+            $data[$property] = $this->{$property};
         }
+
+        //Unpack config objects
+        array_walk_recursive($data, function(&$value, $key)
+        {
+            if($value instanceof KObjectConfigInterface) {
+                $value = KObjectConfig::unbox($value);
+            }
+        });
 
         return $data;
-    }
-
-    public function __debugInfo()
-    {
-        $properties = $this->toArray();
-
-        foreach($properties as $key => $property)
-        {
-            if(is_object($property))
-            {
-                if(method_exists($property, '__debugInfo'))
-                {
-                    ob_start();
-                    var_dump($property);
-                    $debug_info = ob_get_contents();
-                    ob_end_clean();
-
-                    $properties[$key] = $debug_info;
-
-                } elseif(method_exists($property, '__toString')) {
-                    $properties[$key] = (string) $property;
-                }
-                else
-                {
-                    if($property instanceof KObjectInterface)
-                    {
-                        $identifier = (string) $property->getIdentifier();
-                        $properties[$key] = $identifier.' :: ('.  get_class($property) .')';
-                    }
-                    else $properties[$key] = get_class($property);
-                }
-            }
-        }
-
-        return $properties;
-
     }
 
     public function __call($method, $arguments)
